# Office-management-railway-deployment/swagger.yaml
openapi: 3.0.0
info:
  title: Auto Computation Office Management API
  version: 3.0.0
  description: API documentation for the Office Management System Backend
servers:
  - url: http://localhost:3000
    description: Local server
  - url: https://office-management-sbuh.onrender.com
    description: Production server (Render)
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in a cookie
security:
  - cookieAuth: []
paths:
...
  /admin/emp/all:
    get:
      summary: Get all active employees
      tags:
        - Admin-Employees
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of active employees
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
        "500":
          description: Server error
...
  /admin/dashboard/stats:
    get:
      summary: Get administrative dashboard statistics
      tags:
        - Admin-Dashboard
      security:
        - cookieAuth: []
      responses:
        "200":
          description: >-
            Dashboard statistics including staff growth, payroll cost, and
            attendance
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_employees:
                    type: integer
                  employee_growth:
                    type: string
                  payroll_cost:
                    type: number
                  attendance_percentage:
                    type: number
                  attendance_change:
                    type: number
                  payroll_history:
                    type: array
                    items:
                      type: object
                  pending_leaves:
                    type: array
                    items:
                      type: object
                  upcoming_holidays:
                    type: array
                    items:
                      type: object
        "500":
          description: Server error
...
  /employee/dashboard/stats:
    get:
      summary: Get employee personal dashboard statistics
      tags:
        - Employee-Dashboard
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Dashboard data including tasks, attendance, and today's schedule
        "500":
          description: Server error
  /employee/dashboard/clock-in:
    post:
      summary: Clock-in for the day
      tags:
        - Employee-Dashboard
      security:
        - cookieAuth: []
      responses:
        "201":
          description: Clock-in successful
        "403":
          description: IP not authorized
        "409":
          description: Already clocked in
        "500":
          description: Server error
  /employee/dashboard/clock-out:
    post:
      summary: Clock-out for the day
      tags:
        - Employee-Dashboard
      security:
        - cookieAuth: []
      responses:
        "201":
          description: Clock-out successful
        "404":
          description: No clock-in record found
        "409":
          description: Already clocked out
        "500":
          description: Server error
  /employee/dashboard/meetings:
    get:
      summary: Get all meetings assigned to the employee
      tags:
        - Employee-Dashboard
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of meetings
        "500":
          description: Server error
...
  /employee/notifications/all:
    get:
      summary: Get all notifications for the current employee
      tags:
        - Employee-Notifications
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of notifications
        "401":
          description: Unauthorized
        "500":
          description: Server error
  /employee/notifications/mark-read/all:
    patch:
      summary: Mark all notifications as read
      tags:
        - Employee-Notifications
      security:
        - cookieAuth: []
      responses:
        "200":
          description: All notifications marked as read
        "401":
          description: Unauthorized
        "500":
          description: Server error
  /employee/notifications/mark-read/{id}:
    patch:
      summary: Mark a specific notification as read
      tags:
        - Employee-Notifications
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
        "401":
          description: Unauthorized
        "404":
          description: Notification not found
        "500":
          description: Server error
  /employee/notifications/delete/{id}:
    delete:
      summary: Delete a specific notification
      tags:
        - Employee-Notifications
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification deleted successfully
        "401":
          description: Unauthorized
        "404":
          description: Notification not found
        "500":
          description: Server error
...